requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sympy==1.12
pydantic==2.4.2
google-genai==0.3.0
python-dotenv==1.0.0
